+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.5.0 (Oct 07th 2025)
   
 - Enhance usability with automatic summary logging during entity creation with tree-structured entity summary output showing properties, operations, and CRUD mappings
+- Update query.select() method to support filtering instead of creating filters externally.
 
 # 0.4.0 (Sep 26th 2025)